Remove from lobby in game.js
It seemed like adding a "clear" method to the lobby module would be a
good way to separate concerns.  Maybe it was, but it also added several
extra loops, a callback function that required specific arguments, and a
new dedicated method, all of which could be replaced with 15 characters
in game.js.